GC-630: EM Form Errors
This commit adds proper Error/Success Message markup
following the NPR `.form__message` pattern.

For this to work we've had to hack around the
core code quite a bit :-(

Unfortunately, we could not do everything in PHP.
We had to change the JS in bookingsform.js to fix
the Success message.

Frustratingly, the plugin adds minified JS code,
but does NOT use it.

Also: the Grunt build process is broken, so we
ONLY changed the unminified bookingsform.js file.

Functionality also depends on CSS from NPR
and further fixes in the core code of the
Events Manager PRO Overrides plugin..